// ExpenseTrackerUI/generated-sources/openapi/models/ExpenseCategoryDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Financial API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const ExpenseCategoryDto = {
    Drinks: 'Drinks',
    Food: 'Food',
    Groceries: 'Groceries',
    Travel: 'Travel',
    Bills: 'Bills',
    Credit: 'Credit',
    Shopping: 'Shopping'
} as const;
export type ExpenseCategoryDto = typeof ExpenseCategoryDto[keyof typeof ExpenseCategoryDto];


export function ExpenseCategoryDtoFromJSON(json: any): ExpenseCategoryDto {
    return ExpenseCategoryDtoFromJSONTyped(json, false);
}

export function ExpenseCategoryDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): ExpenseCategoryDto {
    return json as ExpenseCategoryDto;
}

export function ExpenseCategoryDtoToJSON(value?: ExpenseCategoryDto | null): any {
    return value as any;
}