// ExpenseTrackerUI/generated-sources/openapi/models/IncomeCategoryDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Financial API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const IncomeCategoryDto = {
    Salary: 'Salary',
    Rent: 'Rent'
} as const;
export type IncomeCategoryDto = typeof IncomeCategoryDto[keyof typeof IncomeCategoryDto];


export function IncomeCategoryDtoFromJSON(json: any): IncomeCategoryDto {
    return IncomeCategoryDtoFromJSONTyped(json, false);
}

export function IncomeCategoryDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): IncomeCategoryDto {
    return json as IncomeCategoryDto;
}

export function IncomeCategoryDtoToJSON(value?: IncomeCategoryDto | null): any {
    return value as any;
}